Code/backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
pydantic-settings

motor
python-dotenv

pytest-asyncio
httpx
firebase-admin
pydantic[email]